+++ src/health/health.controller.spec.ts
@@ -8,7 +8,7 @@ import { HealthController } from './health.controller';
 describe('HealthController', () => {
    let controller: HealthController;
 
-   beforeEach(async () => {
+   beforeAll(async () => {
       const module: TestingModule = await Test.createTestingModule({
          //TODO:Prisma ModuleはGlobalなのに定義しないとエラーが出るので原因を時間があるときに調べる
          imports: [TerminusModule, PrismaModule, HttpModule],
