+++ src/jhs/dto/jhs.dto.ts
