+++ src/prisma/prisma.service.spec.ts
@@ -4,7 +4,7 @@ import { PrismaService } from './prisma.service';
 describe('PrismaService', () => {
    let service: PrismaService;
 
-   beforeEach(async () => {
+   beforeAll(async () => {
       const module: TestingModule = await Test.createTestingModule({
          providers: [PrismaService],
       }).compile();
