+++ src/student/student.service.spec.ts
@@ -35,7 +35,7 @@ describe('StudentService', () => {
    let service: StudentService;
    let prisma: PrismaService;
 
-   beforeEach(async () => {
+   beforeAll(async () => {
       const module: TestingModule = await Test.createTestingModule({
          providers: [
             StudentService,
